Extract KeystreamGenerator class
