+++ solitaire.rb
@@ -1,4 +1,7 @@
+require_relative 'keystreamgenerator'
+
 class Solitaire
+
   def initialize(opts={})
     @message           = opts[:message]
     @encrypted_message = opts[:encrypted_message]
@@ -38,10 +41,7 @@ class Solitaire
     :x_added,
     :converted_encrypted_message,
     :subtracted_message_numbers,
-    :deck
   )
-
-  #encrypt
 
   def sanitize_message
     clean_message
@@ -74,15 +74,14 @@ class Solitaire
     @x_added= ary
   end
 
-#keystream------------------------
-
   def convert_keystream_to_numbers
-    @converted_keystream = convert_to_numbers(keystream).compact
+    @converted_keystream ||= convert_to_numbers(keystream).compact
   end
 
   def keystream
     get_keystream_length
-    @keystream_letters ||= generate_keystream
+    generator = KeystreamGenerator.new(keystream_length)
+    @keystream_letters = generator.generate_keystream
   end
 
   def get_keystream_length
@@ -94,134 +93,8 @@ class Solitaire
       end
   end
 
-  def generate_keystream
-    build_deck
-    generate_letters
-  end
-
-  def build_deck
-    new_deck = (1..52).to_a
-    new_deck << 'A'
-    new_deck << 'B'
-    @deck ||= new_deck
-  end
-
   def length_no_spaces(str)
     str.gsub(/\s+/, '').length
-  end
-
-  def generate_letters
-    new_keystream = []
-    letters_added = 0
-
-    until letters_added == keystream_length do
-      new_keystream << cut_deck_and_output_letter
-      new_keystream.last == nil ? letters_added : letters_added += 1
-    end
-
-    new_keystream.compact.join
-  end
-
-  def cut_deck_and_output_letter
-    move_joker_a
-    move_joker_b
-    triple_cut
-    count_cut
-    output_letter
-  end
-
-  def move_joker_a
-    old_position = deck.index('A')
-
-    if old_position == 53
-      new_position = 1
-    else
-      new_position = old_position + 1
-    end
-
-    @deck = deck.insert(new_position, deck.delete_at(old_position))
-  end
-
-  def move_joker_b
-    old_position = deck.index('B')
-
-    if old_position == 53
-      new_position = 2
-    elsif
-      old_position == 52
-      new_position = 1
-    else
-      new_position = old_position + 2
-    end
-
-    @deck = deck.insert(new_position, deck.delete_at(old_position))
-  end
-
-  def triple_cut
-    joker_a_position = deck.index('A')
-    joker_b_position = deck.index('B')
-
-    move_to_top = []
-    move_to_bottom = []
-
-    if joker_a_position > joker_b_position
-      (53-joker_a_position).times do
-        move_to_top << deck.pop
-      end
-      move_to_top.reverse!
-
-      joker_b_position.times do
-        move_to_bottom << deck.shift
-        move_to_bottom
-      end
-    else
-      (53-joker_b_position).times do
-        move_to_top << deck.pop
-      end
-      move_to_top.reverse!
-
-      joker_a_position.times do
-        move_to_bottom << deck.shift
-        move_to_bottom
-      end
-    end
-
-    @deck = (move_to_top.concat(deck)).concat(move_to_bottom)
-  end
-
-  def count_cut
-    cards_to_insert = []
-
-    card_value =
-      if deck.last.is_a?(String)
-        53
-      else
-        deck.last
-      end
-
-    card_value.times do
-      cards_to_insert << deck.shift
-    end
-    @deck = deck.insert(-2,cards_to_insert).flatten
-  end
-
-  def output_letter
-    top_card =
-      if deck.first.is_a?(String)
-        53
-      else
-        deck.first
-      end
-
-    card_to_convert = deck[top_card]
-
-    if card_to_convert.is_a?(String)
-      nil
-    else
-      card_to_convert -= 26 if card_to_convert > 26
-      NUMBER_CONVERTER[card_to_convert]
-
-    end
   end
 
   def convert_message_to_numbers
@@ -256,8 +129,6 @@ class Solitaire
     end
     group(converted.join).join(" ")
   end
-
-  #decrypt------------------------------
 
   def convert_encrypted_message_to_numbers
     @converted_encrypted_message =
